batch update (#210)
* feat: add real random to nondet VM

* feat(modules): move to services stage 1

* feat(modules): move to services stage 2

* feat(modules): move to services stage 3

* feat(modules): move to services stage 4

* chore: fix clippy warnings

* chore: add clippy

* fix: rework retries to loop over all available providers a few times

* chore: update dependencies

* chore: pass CI

* feat(modules): add scripting stage 1

* feat(modules): handle overloaded in lua

* chore: pass CI

* fix: refactor providers to bypass anthropic

* chore: switch from ellipsis to raise NotImplementedError()

* chore: add order-only dependencies

* feat: add cookie to modules

* chore: update schemas

* chore: pass CI

+++ executor/src/ustar.rs
@@ -173,7 +173,7 @@ impl Archive {
                 continue;
             }
 
-            let mut file_size = 0 as usize;
+            let mut file_size = 0_usize;
             for c in file_size_octal.iter().cloned() {
                 if !(b'0'..=b'7').contains(&c) {
                     anyhow::bail!("invalid octal ascii {}", c)
